Fixes issues where extraction queries can't render tables
