+++ client/js/app/utils/ExplorerUtils.js
@@ -207,77 +207,6 @@ module.exports = {
     return params;
   },
 
-  getQueryDataType: function(query){
-    var isInterval = typeof query.interval === "string",
-    isGroupBy = typeof query.group_by === "string" || (query.group_by instanceof Array && query.group_by.length === 1),
-    is2xGroupBy = query.group_by instanceof Array,
-    dataType;
-
-    if (query.analysis_type === "funnel") {
-      dataType = 'cat-ordinal';
-    }
-
-    // metric
-    else if (!isGroupBy && !isInterval) {
-      dataType = 'singular';
-    }
-
-    // group_by, no interval
-    else if (isGroupBy && !isInterval) {
-      dataType = 'categorical';
-    }
-
-    // interval, no group_by
-    else if (isInterval && !isGroupBy) {
-      dataType = 'chronological';
-    }
-
-    // interval, group_by
-    else if (isInterval && isGroupBy) {
-      dataType = 'cat-chronological';
-    }
-
-    // 2x group_by
-    // TODO: research possible dataType options
-    else if (!isInterval && is2xGroupBy) {
-      dataType = 'categorical';
-    }
-
-    // interval, 2x group_by
-    // TODO: research possible dataType options
-    else if (isInterval && is2xGroupBy) {
-      dataType = 'cat-chronological';
-    }
-
-    else if (query.analysis_type === "extraction") {
-      dataType = 'extraction';
-    }
-    else if (query.analysis_type === "select_unique") {
-      dataType = 'nominal';
-    }
-
-    return dataType;
-  },
-
-  getChartTypeOptions: function(query) {
-    var dataTypes = {
-      'singular':           ['metric'],
-      'categorical':        ['piechart', 'barchart', 'columnchart', 'table'],
-      'cat-interval':       ['columnchart', 'barchart', 'table'],
-      'cat-ordinal':        ['barchart', 'columnchart', 'areachart', 'linechart', 'table'],
-      'chronological':      ['areachart', 'linechart', 'table'],
-      'cat-chronological':  ['linechart', 'columnchart', 'barchart', 'areachart'],
-      'nominal':            ['table'],
-      'extraction':         ['table']
-    };
-    var queryDataType = module.exports.getQueryDataType(query);
-    return dataTypes[queryDataType].concat(['JSON']);
-  },
-
-  responseSupportsChartType: function(query, chartType) {
-    return _.includes(module.exports.getChartTypeOptions(query), chartType);
-  },
-
   encodeAttribute: function(attr) {
     return encodeURIComponent(JSON.stringify(attr));
   },
