+++ test/unit/utils/ChartTypeUtilsSpec.js
@@ -1,0 +1,22 @@
+var assert = require('chai').assert;
+var ChartTypeUtils = require('../../../client/js/app/utils/ChartTypeUtils');
+
+describe('utils/ChartTypeUtils', function() {
+  describe('getQueryDataType', function () {
+    it('returns "categorical" if there is a single item group_by array and no interval', function () {
+      var query = {
+        analysis_type: 'count',
+        group_by: ['single.group_by']
+      };
+      assert.strictEqual(ChartTypeUtils.getQueryDataType(query), 'categorical');
+    });
+
+    it('returns "extraction" if query.analysis_type is extraction', function() {
+      var query = {
+        analysis_type: 'extraction'
+      };
+
+      assert.strictEqual(ChartTypeUtils.getQueryDataType(query), 'extraction');
+    });
+  });
+});
